move generic feature gate code from k8s.io/apiserver to k8s.io/component-base
