+++ staging/src/k8s.io/apiserver/pkg/util/feature/BUILD
@@ -1,30 +1,13 @@
 package(default_visibility = ["//visibility:public"])
 
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-    "go_test",
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["feature_gate_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//vendor/github.com/spf13/pflag:go_default_library",
-        "//vendor/github.com/stretchr/testify/assert:go_default_library",
-    ],
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
     srcs = ["feature_gate.go"],
     importmap = "k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/util/feature",
     importpath = "k8s.io/apiserver/pkg/util/feature",
-    deps = [
-        "//vendor/github.com/spf13/pflag:go_default_library",
-        "//vendor/k8s.io/klog:go_default_library",
-    ],
+    deps = ["//staging/src/k8s.io/component-base/featuregate:go_default_library"],
 )
 
 filegroup(
@@ -36,9 +19,6 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//staging/src/k8s.io/apiserver/pkg/util/feature/testing:all-srcs",
-    ],
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
 )
