+++ staging/src/k8s.io/component-base/featuregate/feature_gate_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package feature
+package featuregate
 
 import (
 	"fmt"
