+++ staging/src/k8s.io/component-base/featuregate/testing/BUILD
@@ -2,11 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["feature_gate_testing.go"],
-    importmap = "k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/util/feature/testing",
-    importpath = "k8s.io/apiserver/pkg/util/feature/testing",
+    srcs = ["feature_gate.go"],
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/component-base/featuregate/testing",
+    importpath = "k8s.io/component-base/featuregate/testing",
     visibility = ["//visibility:public"],
-    deps = ["//staging/src/k8s.io/apiserver/pkg/util/feature:go_default_library"],
+    deps = ["//staging/src/k8s.io/component-base/featuregate:go_default_library"],
 )
 
 filegroup(
